requirements.txt
backports.functools-lru-cache==1.5
cntk==2.7
cycler==0.10.0
enum34==1.1.6
h5py==2.9.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
matplotlib==2.2.4
mss==2.0.22
numpy==1.16.5
opencv-python==4.1.0.25
Pillow==6.1.0
pygame==1.9.2
pynput==1.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
scipy==1.2.2
six==1.12.0
